backend/requirements.txt
# Web Framework
fastapi>=0.95.1
uvicorn>=0.22.0

# Database
sqlalchemy>=2.0.9
alembic>=1.10.4
asyncpg>=0.27.0
psycopg2-binary>=2.9.6

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Validation
email-validator>=2.0.0

# Utilities
requests>=2.28.2
pydantic>=1.10.7
